+++ source/components/organisms/index.ts
@@ -1,3 +1,4 @@
 export { default as Step } from "./Step";
 export { default as SummaryList } from "./SummaryList";
 export { default as PrivacyModal } from "./PrivacyModal/PrivacyModal";
+export { default as LoginModal } from "./LoginModal/LoginModal";
